+++ DependencyInjection/Configuration.php
@@ -206,6 +206,10 @@ class Configuration implements ConfigurationInterface
                             ->scalarNode('stale_while_revalidate')->end()
                         ->end()
                     ->end()
+                    ->scalarNode('etag')
+                        ->defaultValue(false)
+                        ->info('Set a simple ETag which is just the md5 hash of the response body')
+                    ->end()
                     ->scalarNode('last_modified')
                         ->validate()
                             ->ifTrue(function ($v) {